requirements.txt
transformers==4.36.2
torch==2.1.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.6
httpx==0.27.0
jinja2==3.1.3
slowapi==0.1.9
loguru==0.7.2 